backend/requirements.txt
fastapi
uvicorn[standard]
jobspy
pandas
python-dateutil
